requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
decorator==4.3.0
defusedxml==0.7.1
Django==3.2.8
django-cors-headers==3.10.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
et-xmlfile==1.1.0
gensim==4.1.2
idna==3.3
igraph==0.9.8
itypes==1.2.0
Jinja2==3.0.2
joblib==1.1.0
keras==2.7.0
MarkupSafe==2.0.1
networkx==2.6.3
nltk==3.6.5
node2vec==0.4.4
numpy==1.21.2
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.3.4
pycparser==2.20
PyJWT==2.2.0
python-dateutil==2.8.2
python-igraph==0.9.7
python3-openid==3.2.0
pytz==2021.3
regex==2021.10.8
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
texttable==1.6.4
threadpoolctl==3.0.0
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.7
